Revert "add spring security dependency and validation on service"
This reverts commit d81c6d35a0dcc7a84ac48eb88ed4e9caf905a15b.

# Conflicts:
#	src/main/java/com/euflausino/desafiotodo/mapper/TodoMapper.java
#	src/main/java/com/euflausino/desafiotodo/service/TodoService.java
#	target/classes/com/euflausino/desafiotodo/controller/TodoController.class
#	target/classes/com/euflausino/desafiotodo/exception/GlobalExceptionHandler.class
#	target/classes/com/euflausino/desafiotodo/mapper/TodoMapper.class
#	target/classes/com/euflausino/desafiotodo/service/TodoService.class

+++ src/main/java/com/euflausino/desafiotodo/service/TodoService.java
@@ -1,14 +1,13 @@
-package com.euflausino.desafiotodo.service.todo;
+package com.euflausino.desafiotodo.service;
 
-import com.euflausino.desafiotodo.dto.todo.AtualizaTodoRequestDTO;
-import com.euflausino.desafiotodo.dto.todo.TodoRequestDTO;
-import com.euflausino.desafiotodo.dto.todo.TodoResponseDTO;
-import com.euflausino.desafiotodo.entity.todo.Todo;
-import com.euflausino.desafiotodo.exception.todo.TodoNaoEncontradaException;
-import com.euflausino.desafiotodo.mapper.todo.TodoMapper;
-import com.euflausino.desafiotodo.repository.todo.TodoRepository;
+import com.euflausino.desafiotodo.dto.AtualizaTodoRequestDTO;
+import com.euflausino.desafiotodo.dto.TodoRequestDTO;
+import com.euflausino.desafiotodo.dto.TodoResponseDTO;
+import com.euflausino.desafiotodo.entity.Todo;
+import com.euflausino.desafiotodo.exception.TodoNaoEncontradaException;
+import com.euflausino.desafiotodo.mapper.TodoMapper;
+import com.euflausino.desafiotodo.repository.TodoRepository;
 import jakarta.transaction.Transactional;
-import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -30,20 +29,24 @@ public class TodoService {
         return TodoMapper.todoResponseDTO(todo);
     }
 
-    public Page<TodoResponseDTO> listarTodos() {
-        return  TodoMapper.toPageResponse(todoRepository.listarTodosComTrueNaFrente());
+    public List<TodoResponseDTO> listarTodos() {
+        return  TodoMapper.todoResponseListDTO(todoRepository.listarTodosComTrueNaFrente());
     }
 
     @Transactional
-    public Page<TodoResponseDTO> excluir(Long id) {
-        validarTodo(id);
+    public List<TodoResponseDTO> excluir(Long id) {
+        if (!todoRepository.existsById(id)) {
+            throw new TodoNaoEncontradaException("Usuário não encontrado.");
+        }
         todoRepository.deleteById(id);
         return listarTodos();
     }
 
     @Transactional
     public TodoResponseDTO editarTodo(Long id, AtualizaTodoRequestDTO todoRequestDTO) {
-        Todo todo = validarTodo(id);
+        Todo todo = todoRepository.findById(id)
+                .orElseThrow(() -> new TodoNaoEncontradaException("Todo não encontrada."));
+
         if(todoRequestDTO.nome() != null) todo.setNome(todoRequestDTO.nome());
         if(todoRequestDTO.descricao() != null) todo.setDescricao(todoRequestDTO.descricao());
         if(todoRequestDTO.prioridade() != null) todo.setPrioridade(todoRequestDTO.prioridade());
@@ -53,14 +56,8 @@ public class TodoService {
     }
 
     public TodoResponseDTO getTodo(Long id) {
-        Todo todo = validarTodo(id);
-        return TodoMapper.todoResponseDTO(todo);
-    }
-
-    private Todo validarTodo(Long id) {
-       return todoRepository.findById(id)
-                .orElseThrow(() -> new TodoNaoEncontradaException("Todo não encontrada."));
-
+        todoRepository.findById(id).orElseThrow(() -> new TodoNaoEncontradaException("Todo não encontrada."));
+        return TodoMapper.todoResponseDTO(todoRepository.findById(id).get());
     }
 
 }
